pyproject.toml
[tool.poetry]
name = "amorphobot"
version = "0.1.0"
description = ""
authors = ["dsalazarCazanhas"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-telegram-bot = "^21.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
